Lower wmma_gfx11 recipe to amdgpu.wmma
The first cut of the wmma_gfx11 recipe emitted `m`/`n`/`k` as `i64`
attributes and threaded `arch`/`wave_size` through `dynamic_attrs`. Both
are wrong for the upstream `amdgpu.wmma` op:

* `m`/`n`/`k` are declared as confined `i32` attributes — the verifier
  rejects an `i64` literal even when the value is in range.
* `arch`/`wave_size` are not part of the op's signature; they ride on the
  `hc.call_intrinsic` purely for dispatch (recipes are already keyed by
  target) and have no business landing on the created op.

Fix the recipe and give the recipe API the missing primitive: width-
annotated integer literals via `t.i32(...)` / `t.i64(...)` builder
helpers (`TypedIntAttr` under the hood). Plain Python `int`s still
default to `i64` for backwards compatibility.

Operand shape validation falls out for free: the recipe forwards
`a_frag`/`b_frag`/`acc_frag` straight into `amdgpu.wmma`, whose verifier
checks `vector<16xf16>`/`vector<8xf32>` shape constraints when the
created op materializes. Simulator-time `arch=gfx11`/`wave_size=32`
checks already live in `_verify_wmma`.

+++ test/HC/transform-ops.mlir
@@ -13,7 +13,7 @@ module attributes {transform.with_named_sequence} {
     %result_type = transform.hc.get_intrinsic_result_type %call {index = 0 : i64} : (!transform.any_op) -> !transform.type
     %arch = transform.hc.get_intrinsic_attr %call {name = "arch"} : (!transform.any_op) -> !transform.any_param
     %wave = transform.hc.get_intrinsic_attr %call {name = "wave_size"} : (!transform.any_op) -> !transform.any_param
-    %created = transform.hc.create_op "amdgpu.wmma" at %call (%a, %b, %acc) result_types(%result_type) dynamic_attrs ["arch", "wave_size"](%arch, %wave) static_attrs = {k = 16 : i64, m = 16 : i64, n = 16 : i64} : (!transform.any_op, !transform.any_value, !transform.any_value, !transform.any_value, !transform.type, !transform.any_param, !transform.any_param) -> (!transform.any_value)
+    %created = transform.hc.create_op "amdgpu.wmma" at %call (%a, %b, %acc) result_types(%result_type) dynamic_attrs ["arch", "wave_size"](%arch, %wave) static_attrs = {k = 16 : i32, m = 16 : i32, n = 16 : i32} : (!transform.any_op, !transform.any_value, !transform.any_value, !transform.any_value, !transform.type, !transform.any_param, !transform.any_param) -> (!transform.any_value)
     transform.hc.replace_intrinsic_call %call with %created : (!transform.any_op, !transform.any_value) -> ()
     transform.yield
   }
